+++ gradient/cli/experiments.py
@@ -3,14 +3,14 @@ import functools
 
 import click
 
-import gradient.api_sdk.clients.sdk_client
-from gradient import config, constants, exceptions, logger
-from gradient.api_sdk.clients import http_client, sdk_client
+from gradient import constants, utils, api_sdk, exceptions, logger
+from gradient.api_sdk.clients import http_client
 from gradient.cli import common
 from gradient.cli.cli import cli
 from gradient.cli.cli_types import json_string, ChoiceType
-from gradient.cli.common import api_key_option, ClickGroup
+from gradient.cli.common import api_key_option, ClickGroup, deprecated
 from gradient.commands import experiments as experiments_commands
+from gradient.config import config
 
 MULTI_NODE_EXPERIMENT_TYPES_MAP = collections.OrderedDict(
     (
@@ -238,28 +238,36 @@ def common_experiments_create_single_node_options(f):
     return functools.reduce(lambda x, opt: opt(x), reversed(options), f)
 
 
+@deprecated("DeprecatedWarning: \nWARNING: --workspaceUrl and --workspaceArchive "
+            "options will not be included in version 0.6.0")
 @create_experiment.command(name="multinode", help="Create multi node experiment")
 @common_experiments_create_options
 @common_experiment_create_multi_node_options
 def create_multi_node(api_key, **kwargs):
+    utils.validate_workspace_input(kwargs)
     common.del_if_value_is_none(kwargs, del_all_falsy=True)
 
-    sdk_client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
-    command = experiments_commands.CreateMultiNodeExperimentCommand(sdk_client=sdk_client)
+    client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    command = experiments_commands.CreateMultiNodeExperimentCommand(sdk_client=client)
     command.execute(kwargs)
 
 
+@deprecated("DeprecatedWarning: \nWARNING: --workspaceUrl and --workspaceArchive "
+            "options will not be included in version 0.6.0")
 @create_experiment.command(name="singlenode", help="Create single node experiment")
 @common_experiments_create_options
 @common_experiments_create_single_node_options
 def create_single_node(api_key, **kwargs):
+    utils.validate_workspace_input(kwargs)
     common.del_if_value_is_none(kwargs, del_all_falsy=True)
 
-    sdk_client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    sdk_client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
     command = experiments_commands.CreateSingleNodeExperimentCommand(sdk_client=sdk_client)
     command.execute(kwargs)
 
 
+@deprecated("DeprecatedWarning: \nWARNING: --workspaceUrl and --workspaceArchive "
+            "options will not be included in version 0.6.0")
 @create_and_start_experiment.command(name="multinode", help="Create and start new multi node experiment")
 @common_experiments_create_options
 @common_experiment_create_multi_node_options
@@ -273,13 +281,18 @@ def create_single_node(api_key, **kwargs):
 )
 @click.pass_context
 def create_and_start_multi_node(ctx, api_key, show_logs, **kwargs):
-    sdk_client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    utils.validate_workspace_input(kwargs)
+    common.del_if_value_is_none(kwargs, del_all_falsy=True)
+
+    sdk_client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
     command = experiments_commands.CreateAndStartMultiNodeExperimentCommand(sdk_client=sdk_client)
     experiment = command.execute(kwargs)
     if experiment and show_logs:
         ctx.invoke(list_logs, experiment_id=experiment["handle"], line=0, limit=100, follow=True, api_key=api_key)
 
 
+@deprecated("DeprecatedWarning: \nWARNING: --workspaceUrl and --workspaceArchive "
+            "options will not be included in version 0.6.0")
 @create_and_start_experiment.command(name="singlenode", help="Create and start new single node experiment")
 @common_experiments_create_options
 @common_experiments_create_single_node_options
@@ -293,7 +306,10 @@ def create_and_start_multi_node(ctx, api_key, show_logs, **kwargs):
 )
 @click.pass_context
 def create_and_start_single_node(ctx, api_key, show_logs, **kwargs):
-    sdk_client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    utils.validate_workspace_input(kwargs)
+    common.del_if_value_is_none(kwargs, del_all_falsy=True)
+
+    sdk_client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
     command = experiments_commands.CreateAndStartSingleNodeExperimentCommand(sdk_client=sdk_client)
     experiment = command.execute(kwargs)
     if experiment and show_logs:
@@ -311,7 +327,7 @@ def create_and_start_single_node(ctx, api_key, show_logs, **kwargs):
 )
 @click.pass_context
 def start_experiment(ctx, experiment_id, show_logs, api_key):
-    client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
     experiments_commands.start_experiment(experiment_id, client=client)
     if show_logs:
         ctx.invoke(list_logs, experiment_id=experiment_id, line=0, limit=100, follow=True, api_key=api_key)
@@ -321,7 +337,7 @@ def start_experiment(ctx, experiment_id, show_logs, api_key):
 @click.argument("experiment-id")
 @api_key_option
 def stop_experiment(experiment_id, api_key):
-    client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
     experiments_commands.stop_experiment(experiment_id, client=client)
 
 
@@ -329,7 +345,7 @@ def stop_experiment(experiment_id, api_key):
 @click.option("--projectId", "-p", "project_ids", multiple=True)
 @api_key_option
 def list_experiments(project_ids, api_key):
-    client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
     command = experiments_commands.ListExperimentsCommand(sdk_client=client)
     try:
         command.execute(project_id=project_ids)
@@ -341,7 +357,7 @@ def list_experiments(project_ids, api_key):
 @click.argument("experiment-id")
 @api_key_option
 def get_experiment_details(experiment_id, api_key):
-    client = gradient.api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
+    client = api_sdk.clients.sdk_client.SdkClient(api_key=api_key)
     command = experiments_commands.GetExperimentCommand(sdk_client=client)
     try:
         command.execute(experiment_id=experiment_id)