+++ tests/unit/test_workspace.py
@@ -4,7 +4,7 @@ import click
 import mock
 import pytest
 
-from gradient import exceptions
+from gradient import exceptions, utils
 from gradient.workspace import S3WorkspaceHandler
 
 MOCK_BUCKET_NAME = 'bucket_name'
@@ -43,12 +43,14 @@ class TestWorkspace(object):
         with pytest.raises(click.UsageError):
             workspace_handler.handle(params)
 
-    def test_dont_upload_if_archive_url_provided(self, workspace_handler):
+    @mock.patch("gradient.utils.PathParser.parse_path", return_value=utils.PathParser.LOCAL_FILE)
+    def test_dont_upload_if_archive_path_provided(self, _, workspace_handler):
         workspace_handler.handle({'workspaceUrl': 'foo'})
 
         workspace_handler._upload.assert_not_called()
 
-    def test_zip_files_and_receive_s3_response_when_no_dir_provided(self, workspace_handler):
+    @mock.patch("gradient.utils.PathParser.parse_path", return_value=None)
+    def test_zip_files_and_receive_s3_response_when_no_dir_provided(self, _, workspace_handler):
         archive_name = 'foo.zip'
 
         workspace_handler._zip_workspace = mock.MagicMock()
@@ -61,7 +63,8 @@ class TestWorkspace(object):
         workspace_handler._upload.assert_called_with(archive_name, mock_upload_data)
         assert response_url == 's3://{}/{}'.format(MOCK_BUCKET_NAME, MOCK_OBJECT_KEY)
 
-    def test_zip_files_and_receive_s3_response_when_workspace_dir_provided(self, workspace_handler):
+    @mock.patch("gradient.utils.PathParser.parse_path", return_value=utils.PathParser.LOCAL_DIR)
+    def test_zip_files_and_receive_s3_response_when_workspace_dir_provided(self, _, workspace_handler):
         archive_name = 'foo.zip'
 
         workspace_handler._zip_workspace = mock.MagicMock()
@@ -74,10 +77,35 @@ class TestWorkspace(object):
         workspace_handler._upload.assert_called_with(archive_name, mock_upload_data)
         assert response_url == 's3://{}/{}'.format(MOCK_BUCKET_NAME, MOCK_OBJECT_KEY)
 
-    def test_dont_zip_files_and_receive_s3_response_when_workspace_archive_provided(self, workspace_handler):
+    @mock.patch("gradient.utils.PathParser.parse_path", return_value=utils.PathParser.LOCAL_FILE)
+    def test_dont_zip_files_and_receive_s3_response_when_workspace_archive_provided(self, _, workspace_handler):
         workspace_handler._zip_workspace = mock.MagicMock()
 
         response_url = workspace_handler.handle({'projectHandle': 'foo', 'workspaceArchive': 'foo.zip'})
+
+        workspace_handler._zip_workspace.assert_not_called()
+        workspace_handler._upload.assert_called_once()
+        workspace_handler._upload.assert_called_with(os.path.abspath('foo.zip'), mock_upload_data)
+        assert response_url == 's3://{}/{}'.format(MOCK_BUCKET_NAME, MOCK_OBJECT_KEY)
+
+    @mock.patch("gradient.utils.PathParser.parse_path", return_value=utils.PathParser.LOCAL_FILE)
+    def test_dont_zip_files_and_receive_s3_response_when_workspace_archive_provided_with_workspace(self, _,
+                                                                                                   workspace_handler):
+        workspace_handler._zip_workspace = mock.MagicMock()
+
+        response_url = workspace_handler.handle({'projectHandle': 'foo', 'workspace': 'foo.zip'})
+
+        workspace_handler._zip_workspace.assert_not_called()
+        workspace_handler._upload.assert_called_once()
+        workspace_handler._upload.assert_called_with(os.path.abspath('foo.zip'), mock_upload_data)
+        assert response_url == 's3://{}/{}'.format(MOCK_BUCKET_NAME, MOCK_OBJECT_KEY)
+
+    @mock.patch("gradient.utils.PathParser.parse_path", return_value=utils.PathParser.LOCAL_FILE)
+    def test_dont_zip_files_and_receive_s3_response_when_workspace_archive_provided_with_workspace(self, _,
+                                                                                                   workspace_handler):
+        workspace_handler._zip_workspace = mock.MagicMock()
+
+        response_url = workspace_handler.handle({'projectHandle': 'foo', 'workspace': 'foo.zip'})
 
         workspace_handler._zip_workspace.assert_not_called()
         workspace_handler._upload.assert_called_once()