+++ releasenotes.md
@@ -1,17 +1,22 @@
 # Gradient by Paperspace Release Notes
 
+# Release Notes for 0.2.2a3
+### New features
+* deprecated `--workspaceArchive` and `--workspaceUrl` for `experiments create/run`, `run` and `jobs create` commands
+* single `--workspace` option for defining git, S3 and local workspaces
+
 # Release Notes for 0.2.2a2
-# New features
+###New features
 * experimental wizard for creating projects
 ### Fixes
 * fix autocomplete executing prompt for password when double-tab on gradient apiKey
 
 # Release Notes for 0.2.2a1
-# New features
+### New features
 * add looking for new client version
 
 # Release Notes for 0.2.2a0
-# New features
+### New features
 * added `projects create` command
 * added spinner for some long running commands